+++ src/components/Carousel/index.tsx
@@ -3,8 +3,6 @@ import * as React from 'react'
 import { default as Carousel, IProps as ICarouselProps } from './components/Carousel'
 import { default as Slide, IProps as ISlideProps } from './components/Slide'
 
-import './style.css'
-
 interface ICarouselDecorator extends React.ComponentClass<ICarouselProps> {
   Slide: ISlideDecorator
 }