+++ CHANGES.md
@@ -5,3 +5,8 @@ React Carousel
 # 1.0.1
 
 Fixed: make the project as library
+
+# 1.0.2
+
+Fixed: css will no longer overload other styles
+Fixed: removed React from bundle